+++ example/lib/simple_example/my_player.dart
@@ -12,7 +12,7 @@ import 'package:example/shared/util/player_sprite_sheet.dart';
 ///
 /// Rafaelbarbosatec
 /// on 19/10/21
-class MyPlayer extends SimplePlayer with BlockMovementCollision, UseBarLife {
+class MyPlayer extends SimplePlayer with BlockMovementCollision {
   MyPlayer(Vector2 position)
       : super(
           animation: PlayerSpriteSheet.simpleDirectionAnimation,
@@ -25,7 +25,7 @@ class MyPlayer extends SimplePlayer with BlockMovementCollision, UseBarLife {
   Future<void> onLoad() {
     add(
       RectangleHitbox(
-        size: size / 2,
+        size: size / 4,
         position: size / 4,
       ),
     );